Commit for github issue #37.  Plan visualization utilities.
* Added a new option "export_json". If set to true, it will print
  out a json file that is used for a simple UI to visualize demand
  plan and resource utilization (gantt chart)

* Added a new function, exportJSON() to the JobShop class to export
  required data for visualization in a browser.

* Added a new utils directory with 3 files, jobshop.html, jobshop.css,
  jobshop.js.  Using this utility, the user can import a JSON file
  created using the "export_json" option and visualize the plan.

* Two views are available - resource utilization view (Gantt Chart)
  and a demand plan view.

Other minor changes:
-------------------

* Fixed wrong return type for getPriority() function of CalendarShift
  class.

* Added additional getter function, getPlanDate() to Demand class.

* Modified LOGDATA function in the JobShop class to accept a Path argument
  and any arbitrary message

* Removed the DATA_QUALITY enum since it was unnecessary overdesign

+++ sources/com/ushnisha/JobShop/CalendarShift.java
@@ -63,7 +63,7 @@ public class CalendarShift {
      * Return the priority of the shift
      * @return priority of the shift
      */
-    public double getPriority() {
+    public int getPriority() {
         return this.priority;
     }
